requirements.txt
numpy~=1.22.3
PyYAML~=6.0
fastapi~=0.75.1
pydantic~=1.9.0
pandas~=1.4.2
scikit-learn~=1.0.2
FLAML~=1.0.0